terraform-provider-random: Import terraform-provider-random-3.1.0
The "random" provider allows the use of randomness within Terraform
configurations. This is a logical provider, which means that it works entirely
within Terraform's logic, and doesn't interact with any other services.

Based on net/terraform-provider-random, updated to latest stable version and
adjusted for wip/terraform (and, a good excuse to also test go-package-s
Terraform providers).