Add more bits for i586 build of rust on NetBSD.
The openssl-sys package needed a patch, and that needed an
updated checksum.  Unfortunately, can so far not cross-build
the i586 bootstrap kit due to issues with -latomic, but I got
a bootstrap kit for testing "cross-built" natively on an i686
host, therefore the updates to cross.mk (also comment out one
overlooked entry from earlier...).