import uq
USER REQUEST:
=============
uq is a simple, user-friendly alternative to sort | uniq.

It removes duplicate lines from the output, regardless of the order.
Unlike sort | uniq, uq does not sort entries. This allows uq to operate
on continuous streams as well.

DISCLAIMER:
===========
Package is imported into wip and not into the main branch as it has not been
updated in over 3 years.

It builds fine but issues the followig compilation warning:
-trait objects without an explicit `dyn` are deprecated.