kustomize: Import kustomize-4.1.2 as wip/kustomize
kustomize lets you customize raw, template-free YAML files for multiple
purposes, leaving the original YAML untouched and usable as is.

kustomize targets kubernetes; it understands and can patch kubernetes
style API objects. It's like make, in that what it does is declared in
a file, and it's like sed, in that it emits edited text.