flux2: Import flux2-0.13.1 as wip/flux2
Flux is a tool for keeping Kubernetes clusters in sync with sources of
configuration (like Git repositories), and automating updates to
configuration when there is new code to deploy.

Flux version 2 ("v2") is built from the ground up to use
Kubernetes' API extension system, and to integrate with Prometheus and
other core components of the Kubernetes ecosystem. In version 2, Flux
supports multi-tenancy and support for syncing an arbitrary number of
Git repositories, among other long-requested features.

Flux v2 is constructed with the GitOps Toolkit, a set of composable
APIs and specialized tools for building Continuous Delivery on top of
Kubernetes.