tigervnc: update to 1.11.0.1.20.11.
1.11.0

TigerVNC 1.11.0 is now available. This is a new major release of
TigerVNC, but also a security release. Users that rely on the TLS
feature in the viewers are recommended to upgrade as soon as possible.

Lots of changes have been made since the last release, but the
highlights are:

- A security issue has been fixed in how the viewers handle TLS
certificate exceptions
- vncserver has gotten a major redesign to be compatible with modern
distributions
- The native viewer now has touch gestures to handle certain mouse
actions (e.g. scroll wheel)
- Middle mouse button emulation in the native viewer, for devices
with only two mouse buttons
- The Java viewer now supports Java 9+, but also now requires Java 8+
- Support for alpha cursors in the Java viewer (a feature already
supported in the native viewer)
- The password and username can now be specified via the environment
for the native viewer
- Support for building Xvnc/libvnc.so with Xorg 1.20.7+ and deprecate
support for Xorg older than 1.16
- The official builds have been fixed to work on the upcoming macOS 11
- The Windows server (WinVNC) is now packaged separately as it is
unmaintained and buggy

1.10.1

This is a security release to fix a number of issues that were found by
Kaspersky Lab. These issues affect both the client and server and could
theoretically allow an malicious peer to take control over the software
on the other side.

No working exploit is known at this time, and the issues require the
peer to first be authenticated. We still urge users to upgrade when
possible.

1.10.0

TigerVNC 1.10.0 is now available. Lots of changes have been made since
the last release, but the highlights are:

* The clipboard now supports full Unicode in the native viewer, WinVNC
and Xvnc/libvnc.so
* The native client will now respect the system trust store when
verifying server certificates
* Improved compatibility with VMware's VNC server
* Improved compatibility with some input methods on macOS
* Improvements to the automatic "repair" of JPEG artefacts
* Better handling of the Alt keys in some corner cases
* The Java web server has been removed as applets are no longer support
by most browsers
* x0vncserver can now be configured to only allow local connections
* x0vncserver has received fixes for when only part of the display is shared
* Polling is now default in WinVNC as that works better for most