rust: drop docs from bootstrap kit and speed up install.sh.
When building the bootstrap kit, drop "docs" and "compiler-docs"
from the build.  This will make the install of the bootstrap kit
a lot faster, since it drops 5-6.000 files from the kit, which would
need to be processed by the install.sh script.  Note that this will
also impact the rust-bin package, since we also use the bootstrap
kits for installation there for the ports we maintain locally.  To
get these "docs" sub-sets installed, use the full "lang/rust" normal
build and/or a corresponding binary package.

Also, speed up the install.sh script (via install-template.sh) by
significantly reducing the number of forks in the inner loop.
Instead of relying on "grep", "sed" and "cut" for pattern matching,
substitution and field extraction, use case/esac and parameter
expansion modification to do the same, i.e. using in-shell mechanisms.
Submitted upstream, ref.

https://github.com/rust-lang/rust-installer/pull/111

On an emulated (and loaded...) armv7 system, this reduced the time
(as shown with csh "time") for a test rig of install.sh (which
doesn't even actually do the file copying etc.) from

13000.143u 48616.601s 14:21:27.13 119.2%   57+151k 598+7367io 0pf+0w
to
3247.518u 9754.898s 3:45:48.70 95.9%   31+75k 927+4952io 14pf+0w

(This is with the "docs" and "compiler-docs" still in the bootstrap kit.)

Bump PKGREVISION.