blas64, lapack64, cblas*, lapacke*, openblas*, mk/blas64.bl3: fixup and 64 bit BLAS
This prepares for 64 bit index BLAS libraries alongside 32 bit ones. In preparation
of better support for vendor BLAS libraries, I had to realize that you better
use the C interfaces supplied by them, not the netlib one strapped on. A simple
reason of practicability: The vendor blas libraries, like openblas, like to
ship all symbols in one library, so you get them whether you want it or not.

The netlib reference pacakges are still all separate packages for the separate
libraries, since this is instructive and helps keeping things small when you
do not need all of them. The installation location of the headers is now in
a subdirectory to be able to have 32 and 64 bit variants independent. The 32 bit
ones are linked to ${PREFIX}/include to keep the old picture. We could be brave
and remove those, but there is some value in a build just trying -lcblas and

to be happy.

The blas64.buildlink3.mk conflicts with blas.buildlink3.mk to avoid hazardous
combination of the same symbols with differing ABI. The contents of blas.bl3
will be updated to behave the same way on merging in. One might consider merging
the identical parts of the files, but then, this might be not worth it.

Another idea is to have it all on one blas.bl3 and have a variable to switch
between 32 and 64 bits. BLAS_INTERFACE64=yes, perhaps? BLAS_ILP64 seems to
confuse people a lot;-) That way, a mix of differing settings would also not
be possible.

I would like to merge this before the next branch … let's see. The 64 bit stuff
is pretty safe right now as nothing else uses it, but I intend to make
py-numpy, octave, R etal. use the 64 bit CBLAS on 64 bit platforms.
That switch itself will need some time to settle, I guess, but having the 64 bit
BLAS in main is a prerequisite to do that without too much hassle.