terraform-provider-aws-modules-http: Import terraform-provider-aws-modules-http-2.4.1
The Terraform HTTP provider is a utility provider for interacting with
generic HTTP servers as part of a Terraform configuration.

This provider requires no configuration.

This project is a fork of upstream terraform-provider-http project,
which is not taking time to review the PR as of writing of this. This
project does not replace that nor we intend to maintain it outside of
any future need for internal terraform-aws-eks related improvements
and/or bug fixes.