www/xurl: import package
A little tool to manipulate urls.
Xurl offer severals commands:

-Dissect: The dissect command splits and prints the different parts of the url.
These parts are; scheme, for urls with authentication information, the username
and the password, host, port, path to the document, query string and fragment.
-Join: The join command joins an url and a new path together.
-Encode: The encode command applies the percent-encoding on the url.
-Decode: The decode command reverses the percent-encoding applyied on an url.
-Idna Encode: The idna-encode command applies the punycode translation on an
internationalized domain name.
-Idna Decode: The idna-decode command reverses the punycode translation applied
on an internationalized domain name.

Putting this here until freeze is over and upstream provides tags.