mummer4: import mummer-4.0.0rc1 (4.0.0.20201002) as wip/mummer4
MUMmer is a versatile alignment tool for DNA and protein sequences.

MUMmer is a system for rapidly aligning DNA and protein sequences. The
nucmer aligner in the current version (release 4.x) can align two
mammalian genomes in about 3 hours on a typical 32+ core workstation
with 64+Gb RAM; smaller genomes such as bacteria or small eukaryotes
are aligned in seconds or minutes. The promer utility generates
alignments based upon the six-frame translations of both input
sequences. promer permits the alignment of genomes for which the
proteins are similar but the DNA sequence is too divergent to detect
similarity. See the nucmer and promer readme files in the "docs/"
subdirectory for more details. MUMmer is open source, and we ask that
you cite our most recent paper in any publications that use this
system:

Marcais G, Delcher AL, Phillippy AM, Coston R, Salzberg SL, Zimin
A. MUMmer4: A fast and versatile genome alignment system. PLoS
computational biology. 2018 Jan 26;14(1):e1005944.