terraform-provider-github: Remove commented out UNLIMIT_RESOURCES
That was a copypasto from terraform-provider-local that I have used as a
starting point... hopefully also limited resources will work for this package!
(I have not seen any problem with that but they are usually triggered on bulk
builds... but probably this Terraform provider is relatively "small" unlike e.g.
AWS one)