pdfjam3: import pdfjam3-3.03 as wip/pdfjam3
* Motivation for a separate package:

This updates print/pdfjam to version pdfjam-3.03, The project
has a new maintainer and the scripts pdf90, pdfnup, and so on,
are no longer supported or distributed with the project.  The
(unsupported) scripts are kept in a separate repository:

 https://github.com/rrthomas/pdfjam-extras

To avoid user surprises (i.e. all the commands disappeared),
the package is currently named pdfjam3, and the script and
documentation as well.

* Description:

The pdfjam package makes available the pdfjam shell script that
provides a simple interface to much of the functionality of the
excellent pdfpages package (by Andreas Matthias) for LaTeX. The
pdfjam script takes one or more PDF files (and/or JPG/PNG graphics
files) as input, and produces one or more PDF files as output. It
is useful for joining files together, selecting pages, reducing
several source pages onto one output page, etc., etc.