helm: Update to 3.6.2
Changes:
3.6.2
-----
Helm v3.6.2 is a regression fix release from v3.6.1. The security fix
released in v3.6.1 caused a bug to surface that was previously only an
issue in uncommon situations. v3.6.2 provides a solution to this bug.

3.6.1
-----
Helm v3.6.1 is a security (patch) release. Users are strongly
recommended to update to this release.

While working on the Helm source, a Helm core maintainer discovered a
situation where the username and password credentials associated with a
Helm repository could be passed on to another domain referenced by that
Helm repository. More information can be found in the security
advisory.

Note: if updating to v3.6.1 causes issues downloading charts from
password protected repositories it may mean the charts are being served
from a different domain than the repository. If you trust the other
domain to receive the credentials and want to pass the credentials to
that domain name you can use the new --pass-credentials flag. This flag
works with helm repo add as well as commands that accept the --repo
flag.

3.6.0
-----
Helm v3.6.0 is a feature release. Users are encouraged to upgrade for
the best experience.

Notable changes:
 - Apple silicon support
 - Continued OCI changes as the experiment is refined
 - Improved auto-completion
 - Helm create ingress updates

3.5.4
-----
Helm v3.5.4 is a patch release. Users are encouraged to upgrade for the
best experience.