wk: Import R package wk-0.5.0 as wip/R-wk
The goal of wk is to provide lightweight R, C, and C++ infrastructure
for a distributed ecosystem of packages that operate on collections of
coordinates. First, wk provides vector classes for points, circles,
rectangles, well-known text (WKT), and well-known binary (WKB). Second,
wk provides a C API and set of S3 generics for event-based iteration
over vectors of geometries.