R-triebeard: Import R package triebeard-0.3.0 as wip/R-triebeard
Fast key-value matching in R and Rcpp

Tries, or [radix trees](https://en.wikipedia.org/wiki/Radix_tree), are
key-value data structures optimised for very, very fast matching of
the keys against user-provided data (and then the return of the
associated values!)

This is pretty useful in data cleaning and value extraction, and tries
let you do it *really* efficiently. `triebeard` contains an
implementation that can be used both when writing R, and when writing
Rcpp (and imported and linked against, to boot). For more information
see:

1. The [vignette on Rcpp
   usage](https://cran.r-project.org/web/packages/triebeard/vignettes/rcpp_radix.html);

2. The [vignette on R
   usage](https://cran.r-project.org/web/packages/triebeard/vignettes/r_radix.html).