R-crul: Import of R package crul-1.1.0 as wip/R-crul
A simple HTTP client, with tools for making HTTP requests, and mocking
HTTP requests. The package is built on R6, and takes inspiration from
Ruby's 'faraday' gem (<https://rubygems.org/gems/faraday>). The
package name is a play on curl, the widely used command line tool for
HTTP, and this package is built on top of the R package 'curl', an
interface to 'libcurl' (<https://curl.se/libcurl/>).