(textproc/R-vroom) import package R-vroom-1.5.4
The goal of 'vroom' is to read and write data (like 'csv', 'tsv' and
'fwf') quickly. When reading it uses a quick initial indexing step,
then reads the values lazily , so only the data you actually use needs
to be read.  The writer formats the data in parallel and writes to
disk asynchronously from formatting.