trf: Import trf-4.09.1 as wip/trf.
A tandem repeat in DNA is two or more adjacent, approximate copies of
a pattern of nucleotides. Tandem Repeats Finder is a program to locate
and display tandem repeats in DNA sequences. In order to use the
program, the user submits a sequence in FASTA format. There is no need
to specify the pattern, the size of the pattern or any other
parameter. The output consists of two files: a repeat table file and
an alignment file. The repeat table, viewable in a web browser,
contains information about each repeat, including its location, size,
number of copies and nucleotide content. Clicking on the location
indices for one of the table entries opens a second browser page that
shows an alignment of the copies against a consensus pattern. The
program is very fast, analyzing sequences on the order of .5Mb in just
a few seconds. Submitted sequences may be of arbitrary length. Repeats
with pattern size in the range from 1 to 2000 bases are detected.