pharo-vm: Import pharo-vm-9.0.9 as wip/pharo-vm
This is the branch of the Headless VM used for Pharo. This branch has
the modifications required to run Pharo in a true headless
environment. The image running on the VM is responsible of handling
the UI and the events.

The image includes a default implementation of the handling of UI
through the use of SDL2.

For more details about the whole Pharo VM project refer to our wiki.

This is a fork of OpenSmalltalk-vm. We are doing our best to keep
compatibility and contribute back, as long as it fits the objective of
Pharo community.