CuraEngine: Import CuraEngine-4.11.0 as wip/CuraEngine
The CuraEngine is a C++ console application for 3D printing GCode
generation. It has been made as a better and faster alternative to the
old Skeinforge engine.

The CuraEngine is pure C++ and uses Clipper from
http://www.angusj.com/delphi/clipper.php Furthermore it depends on
libArcus by Ultimaker, which can be found at
http://github.com/Ultimaker/libArcus

This is just a console application for GCode generation. For a full
graphical application look at https://github.com/Ultimaker/Cura which
is the graphical frontend for CuraEngine.

The CuraEngine can be used seperately or in other applications. Feel
free to add it to your application. But please take note of the
License.