routinator: shuffle around setting of MASTER_SITES and GITHUB_TAG.
Still not satisfied, "-v<version>" becomes not really the version
number, but part of the package name due to the "v" presence.
However, trying to get rid of the "v" in various way too easliy
leads to "cannot do 'make fetch'" if you clean out the previously
fetched distfile first.