doomlegacy-devel: Update to SVN revision 1599
This fixes some bad subsector assigns for hardware-draw
in Avactor.wad (Map05 sector 1502).

The function Find_Poly_Sector is used on a polygon, to find the its
sector.
With subsectors that have a seg as a side, this can be done by
inspection.  For subsectors that are not adjacent to a sector side,
Find_Poly_Sector is used to find the sector that they are within.
This is done by finding a linedef of that sector.

Find_Poly_Sector generates a vertex within the poly by averaging the
polygon vertexes.  It then searches all the linedef for the closest.
The calculation for closest overflows sometimes.

When the wrong linedef is deemed to be closest, the sidedef of that
linedef is used to assign the sector to the subsector.
This makes the wrong sector be used to display the floor height, and
floor flat.

I tried several approaches to this.  I ended up using of the
current code, rewritten to be faster, and more robust in large maps.
The linedef considered are limited to those that bracket the test point.
The measurement is the distance to the line, on the same x or y
coordinate.  This ensures that the distance is to a point on the line
that is within the linedef boundaries.  All the distances are measured
along the same lines, so it is the closest linedef along a particular
line.
This still finds a linedef of the enclosing sector because any wrong
linedef would be beaten by the linedef of the sector that crosses that
line.

Using a cross product failed due to it finding the closest point to
the linedef without restricting it to be within the boundaries of the
linedef.
A far linedef that was angled could be considered closest to the
test point, even though it was farther away using other criterea.

Eliminating linedef from test consideration using dx vrs dy had to be
removed.
There may be only one linedef that beats another wrong linedef
from being used.  Every linedef has to be tested for every geometric
relationship.