doomlegacy-devel: Update to SVN revision 1602
Avactor.wad has some masked textures used on upper textures and single
sided walls.  This is not entirely valid usage, but happens.
The software render draws black in the missing portions of the texture.
OpenGL is drawing a masked texture, showing sky through the wall.
Changed the drawing of such textures in hardware draw to not use the
texture alpha.
Changed the SplitWall function to support such alternative drawing
blends.