mtail: Import mtail-3.0.0rc47 as wip/mtail
mtail is a tool for extracting metrics from application logs to be
exported into a timeseries database or timeseries calculator for
alerting and dashboarding.

It fills a monitoring niche by being the glue between applications that
do not export their own internal state (other than via logs) and
existing monitoring systems, such that system operators do not need to
patch those applications to instrument them or writing custom
extraction code for every such application.

The extraction is controlled by mtail programs which define patterns
and actions.

Metrics are exported for scraping by a collector as JSON or Prometheus
format over HTTP, or can be periodically sent to a collectd, StatsD, or
Graphite collector socket.