dangerzone: add incomplete package (PySide2 needs packaging)
Take potentially dangerous PDFs, office documents, or images and
convert them to a safe PDF.

Dangerzone works like this: You give it a document that you don't
know if you can trust (for example, an email attachment). Inside
of a sandbox, Dangerzone converts the document to a PDF (if it
isn't already one), and then converts the PDF into raw pixel data:
a huge list of of RGB color values for each page. Then, in a separate
sandbox, Dangerzone takes this pixel data and converts it back into
a PDF.