1st release to wip of DLB's pkgsrc attempt of Mutabah's Rust Compiler.  [Currently for rust 1.39, only]
Builds are not seamless yet (sometimes requiring repeated
invocations, e.g., of "make update", to get past sporadic
"guard page"--and other possibly memory related--compile-time
errors); and in addition, the packages need to be better tuned for
pkgsrc correctness by a knowledgeable pkgsrc developer.

For example, a big win for troubleshooting might be to separate
out the relatively reliable llvm build stage into its own sub-
package if that's possible; so as to avoid always doing a very
expensive rebuild of that phase when it isn't actually necessary.

Also, there's probably NetBSD bias in the patches & packages; and
it may be possible to improve this to benefit other platforms.

Moreover, these were constructed and tested under a private copy
of pkgsrc, in .../lang , so they may immediately fail and need to
be re-configured for .../wip

That said, with adequate RAM [~12G], swap [~10G], storage [~35G]
and determination, these can successfully complete all the way
through the packaging & installation (on a NetBSD 9/amd64 system)
of rustc & cargo binaries--built from rustc sources, but using the
compiled-from-C++-source mrustc toolchain only (i.e., no rustc
binaries downloaded during the bootstrap).

Invocation of the resulting exe's with --help or --version args
currently works (and the exe's SHOULD also pass ~99% of the rustc
test suite... but I haven't re-run the suite lately).