fetchmail: add a pre-release of version 7.0.
This isn't a "real" alpha10, but the most recent tag on the
"next" branch since alpha9, tag 9c011ca4.

For me, the major difference between this version and the version
in pkgsrc is that this one is supposed to support oauth2 authentication.

It builds, it installs, for me, on NetBSD/amd64 9.99.92.
Not much other testing done yet.

I tried for a while using wip/mk/gitlab.mk and have left in some
comments related to that, but could not make the pieces fit together
so that it tries to fetch the right file.