et: fix build on NetBSD
Fix build on NetBSD by requiring gcc 9.
This removes the linker error when
using std::filesystem.
Since gcc 9 has std::filesystem in stdc++
don't link to stdc++fs on NetBSD.
Telemetry is not supported on NetBSD and
FreeBSD so disable it in pkgsrc to make
the PLIST consistent.
While here depend on cxxopts from pkgsrc
instead of the bundled one.