mk/git-package.mk: Improve check if update of Git archive is necessary
Git archives are supposed to be updated when update of Git working
area causes a change in ref related to the desired revision.
To achieve this, refs before and after update are compared.

Unfortunately the refs related to HEAD were probed around checkout
step, which caused the comparison to always result in a match.

This change moves the probes around fetch step which
actually updates the refs, and makes them check
the desired revision instead of hardcoded HEAD.