Import Mighty Mike 3.0.0 as wip/mightymike
This is Pangea Software's Mighty Mike updated to run on modern systems. Set
in a toy store, this top-down action game is a staple of 90's Macintosh
games. It was initially published in 1995 under the name Power Pete.

Note: it currently does NOT run on my NetBSD/amd64 9.2 -- I'm betting it's
related to one of the patches I've applied which ignores some code that
fails at build-time due to C++ std mismatch